make sure video pauses at correct position
previously it was assumed that one slice is equal to one frame:
// Assume 1 slice per frame

but my 30 second video has 900 frames and 917 slices. if the video was
played with repeat=false, it was paused at about 110% of the video
playback, not the last frame (=100%)

the decode now has sized queues for input packages and output frames.

also renamed some variables to show the difference between packages and
frames, and input and output of the decoder